Fix almost all remaining uses of sealtk
Change most uses of the old sealtk namespace to the new krest. This
should take care of the last vestiges of the old name aside from one use
of a namespace alias, and a comment in a TIFF file that isn't worth
changing.

Additionally, this should get the repository back to a state where it
can be built successfully.