Optionally use fletch
Add logic to optionally attempt to use fletch, the collection of
third-party dependencies which is also used by KWIVER. This may simplify
providing Krest's own dependencies. Also, rearrange find calls to be in
decreasing order of dependence, which should reduce the chances of
finding different versions of dependencies when both Krest and another
dependency need the same package.