Expose Player placeholder text
Rename API to set text displayed by Player when no image is available,
and promote it to a proper, public property. There are a bunch of users
of this, now, and each one seems to want different text, and the old
default was only really suitable for the NOAA GUI; having to subclass
the Player everywhere just to be able to change the placeholder text was
getting annoying.