Fix image to texture conversion
Adjust stride computation in image to texture conversion. We want
strides reported in bytes, and our logic to check for valid strides is
based on the same, however KWIVER reports strides in increments of the
value type. This was fine for 8-bit images and seems to have been
working by accident for single-channel 16-bit images, but was causing
the stride check to incorrectly reject some multi-channel 16-bit images.